Resolve invalid UUID generation
Issue #6

UUIDs were generated using a random set of alphanumeric characters.
This resulted in times when the UUID began with a number. Since the
UUID is used as an element in the server.xml file, when this
occurred, xmlstarlet would fail complaining that the UUID was not
a valid XML element. This patch ensures that the first character of
the generated UUID is always an alpha character. The others should
remain alphanumeric.