Collapse build and cleanup to reduce image size.
This collapses all build and cleanup steps into one RUN directive to
reduce the image size. Docker starts a new overlay layer for every
directive which means that the cleanup is just a "hiding" of the content
to clean in a new layer but not reducing the actual image size.

With this change the image size shrinks from 912MB to 585MB.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>